+++ lib/views/cadernoCampo.dart
@@ -1,0 +1,766 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_safraapp/views/mapsPage.dart';
+
+class cadernoCampoPage extends StatefulWidget {
+  const cadernoCampoPage({super.key});
+
+  @override
+  State<cadernoCampoPage> createState() => _cadernoCampoPageState();
+}
+
+@override
+class _cadernoCampoPageState extends State<cadernoCampoPage> {
+  String? _selectedValue_classe;
+  String? _selectedValue_toxicidade;
+  String? _selectedValue_opcoesmedidas;
+  String? _selectedValue_medidatempo;
+  String? _selectedValue_medidaterreno;
+
+  final List<String> _opcoes_medidas = [
+    '',
+    'L/ha',
+    'Kg/ha',
+    'mL/ha',
+    'g/ha',
+    'mL/L',
+    'g/L',
+    'UI/ha',
+    'gotas/m2',
+  ];
+
+  final List<String> _medidas_tempo = [
+    '',
+    'Hora(s)',
+    'Dia(s)',
+    'Semana(s)',
+    'Mes(es)',
+    'Bimestre(s)',
+    'Trimestre(s)',
+    'Semestre(s)',
+    'Ano(s)',
+  ];
+
+  final List<String> _opcoes_classe_agronomica = [
+    '',
+    'Acaricida',
+    'Fungicida',
+    'Herbicida',
+    'Inseticida',
+    'Nematicida',
+  ];
+
+  final List<String> _opcoes_nivel_toxicologico = [
+    '',
+    'Categoria 1 - Extremamente Tóxico',
+    'Categoria 2 - Altamente Tóxico',
+    'Categoria 3 - Moderadamente Tóxico',
+    'Categoria 4 - Pouco Tóxico',
+    'Categoria 5 - Improvável de Causar Dano Agudo',
+    'Não Classificado',
+  ];
+
+  final List<String> _medidas_terreno = [
+    '',
+    'Hectare(s)',
+    'Alqueire(s) Paulista',
+    'Alqueire(s) Mineiro',
+  ];
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Row(
+          mainAxisAlignment: MainAxisAlignment.start,
+          children: [
+            Text(
+              'Caderno de Campo',
+              style: TextStyle(fontSize: 22),
+            ),
+          ],
+        ),
+        automaticallyImplyLeading: true,
+        flexibleSpace: Center(
+          child: Padding(
+            padding: const EdgeInsets.only(top: 30.0),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.end,
+              children: [
+                Image.asset(
+                  'images/Logo_SafraApp3.png', // Substitua pelo caminho da sua imagem
+                  height: 40, // Ajuste a altura conforme necessário
+                ),
+              ],
+            ),
+          ),
+        ),
+        //automaticallyImplyLeading: false,
+        elevation: 0,
+        backgroundColor: Color.fromARGB(255, 2, 89, 47),
+        centerTitle: true,
+      ),
+      body: SingleChildScrollView(
+        child: Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.only(top: 28.0, left: 12),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.start,
+                children: [
+                  Text(
+                    'Identificação do Talhão',
+                    style: TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
+                  ),
+                ],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: TextField(
+                  //keyboardType: TextInputType.number,
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: 'Nome',
+                      hintStyle:
+                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                ),
+              ),
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                SizedBox(width: 5.5),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 1.52,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          topLeft: Radius.circular(50),
+                          bottomLeft: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Tamanho da Área',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 3.95,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          topRight: Radius.circular(50),
+                          bottomRight: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: DropdownButton(
+                      value: _selectedValue_medidaterreno,
+                      isExpanded: true,
+                      hint: const Text(
+                        'Medida',
+                        style: TextStyle(
+                          color: Color.fromARGB(255, 8, 46, 28),
+                        ),
+                      ),
+                      items: _medidas_terreno.map<DropdownMenuItem<String>>(
+                        (String value) {
+                          return DropdownMenuItem<String>(
+                            value: value,
+                            child: Text(value),
+                          );
+                        },
+                      ).toList(),
+                      onChanged: (String? newValue) {
+                        setState(() {
+                          _selectedValue_medidaterreno = newValue;
+                        });
+                      },
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: TextField(
+                  //keyboardType: TextInputType.number,
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: 'Cultura',
+                      hintStyle:
+                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: TextField(
+                  //keyboardType: TextInputType.number,
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: 'Variedade',
+                      hintStyle:
+                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                ),
+              ),
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 2.25,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.all(
+                          Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Data de Plantio',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+                SizedBox(width: 5.5),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 2.25,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.all(
+                          Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Sistema de Plantio',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 28.0, left: 12),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.start,
+                children: [
+                  Text(
+                    'Mapeamento Geográfico',
+                    style: TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
+                  ),
+                ],
+              ),
+            ),
+            InkWell(
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(builder: (context) => mapsPage()),
+                );
+              },
+              child: Padding(
+                padding: const EdgeInsets.only(top: 18.0),
+                child: Container(
+                  width: MediaQuery.of(context).size.width / 1.1,
+                  height: 300,
+                  decoration: BoxDecoration(
+                      borderRadius: BorderRadius.all(
+                        Radius.circular(20),
+                      ),
+                      border: Border.all(
+                          color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                  child: Center(
+                    child: Text('Google Maps'),
+                  ),
+                ),
+              ),
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 2.25,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          bottomLeft: Radius.circular(50),
+                          topLeft: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Latitude',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+                SizedBox(width: 5.5),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 2.25,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          bottomRight: Radius.circular(50),
+                          topRight: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Longitude',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 28.0, left: 12),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.start,
+                children: [
+                  Text(
+                    'Aplicação',
+                    style: TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
+                  ),
+                ],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: TextField(
+                  //keyboardType: TextInputType.number,
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: 'Nome do Produto',
+                      hintStyle:
+                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: DropdownButton(
+                  value: _selectedValue_classe,
+                  isExpanded: true,
+                  hint: const Text(
+                    'Alvo Biológico',
+                    style: TextStyle(
+                      color: Color.fromARGB(255, 8, 46, 28),
+                    ),
+                  ),
+                  items:
+                      _opcoes_classe_agronomica.map<DropdownMenuItem<String>>(
+                    (String value) {
+                      return DropdownMenuItem<String>(
+                        value: value,
+                        child: Text(value),
+                      );
+                    },
+                  ).toList(),
+                  onChanged: (String? newValue) {
+                    setState(() {
+                      _selectedValue_classe = newValue;
+                    });
+                  },
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: DropdownButton(
+                  value: _selectedValue_toxicidade,
+                  isExpanded: true,
+                  hint: const Text(
+                    'Nível de Toxicidade',
+                    style: TextStyle(
+                      color: Color.fromARGB(255, 8, 46, 28),
+                    ),
+                  ),
+                  items: _opcoes_nivel_toxicologico
+                      .map<DropdownMenuItem<String>>((String value) {
+                    return DropdownMenuItem<String>(
+                      value: value,
+                      child: Text(value),
+                    );
+                  }).toList(),
+                  onChanged: (String? newValue) {
+                    setState(() {
+                      _selectedValue_toxicidade = newValue;
+                    });
+                  },
+                ),
+              ),
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 2.95,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.all(
+                          Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.datetime,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Data da Aplicação',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+                SizedBox(width: 5.5),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 3.25,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          topLeft: Radius.circular(50),
+                          bottomLeft: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Dose Aplicada',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 3.95,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          topRight: Radius.circular(50),
+                          bottomRight: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: DropdownButton(
+                      value: _selectedValue_opcoesmedidas,
+                      isExpanded: true,
+                      hint: const Text(
+                        'Medida',
+                        style: TextStyle(
+                          color: Color.fromARGB(255, 8, 46, 28),
+                        ),
+                      ),
+                      items: _opcoes_medidas.map<DropdownMenuItem<String>>(
+                        (String value) {
+                          return DropdownMenuItem<String>(
+                            value: value,
+                            child: Text(value),
+                          );
+                        },
+                      ).toList(),
+                      onChanged: (String? newValue) {
+                        setState(() {
+                          _selectedValue_opcoesmedidas = newValue;
+                        });
+                      },
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: TextField(
+                  //keyboardType: TextInputType.number,
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: 'Método de Aplicação',
+                      hintStyle:
+                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                ),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 50,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(50),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: TextField(
+                  //keyboardType: TextInputType.number,
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: 'Responsável pela Aplicação',
+                      hintStyle:
+                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                ),
+              ),
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                SizedBox(width: 5.5),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 1.52,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          topLeft: Radius.circular(50),
+                          bottomLeft: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: TextField(
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                          border: InputBorder.none,
+                          hintText: 'Período de Carência',
+                          hintStyle:
+                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                    ),
+                  ),
+                ),
+                Padding(
+                  padding: const EdgeInsets.only(top: 18.0),
+                  child: Container(
+                    width: MediaQuery.of(context).size.width / 3.95,
+                    height: 50,
+                    padding: EdgeInsets.all(8),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.only(
+                          topRight: Radius.circular(50),
+                          bottomRight: Radius.circular(50),
+                        ),
+                        color: Colors.white,
+                        border: Border.all(
+                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                    child: DropdownButton(
+                      value: _selectedValue_medidatempo,
+                      isExpanded: true,
+                      hint: const Text(
+                        'Medida',
+                        style: TextStyle(
+                          color: Color.fromARGB(255, 8, 46, 28),
+                        ),
+                      ),
+                      items: _medidas_tempo.map<DropdownMenuItem<String>>(
+                        (String value) {
+                          return DropdownMenuItem<String>(
+                            value: value,
+                            child: Text(value),
+                          );
+                        },
+                      ).toList(),
+                      onChanged: (String? newValue) {
+                        setState(() {
+                          _selectedValue_medidatempo = newValue;
+                        });
+                      },
+                    ),
+                  ),
+                ),
+              ],
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 18.0),
+              child: Container(
+                width: MediaQuery.of(context).size.width / 1.1,
+                height: 200,
+                padding: EdgeInsets.all(8),
+                decoration: BoxDecoration(
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(20),
+                    ),
+                    color: Colors.white,
+                    border: Border.all(
+                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
+                child: TextField(
+                  //keyboardType: TextInputType.number,
+                  decoration: InputDecoration(
+                      border: InputBorder.none,
+                      hintText: 'Observações',
+                      hintStyle:
+                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
+                ),
+              ),
+            ),
+            Container(
+              padding: EdgeInsets.only(top: 50, bottom: 10),
+              child: InkWell(
+                onTap: () {},
+                child: Container(
+                  height: 60,
+                  width: MediaQuery.of(context).size.width / 1.1,
+                  decoration: BoxDecoration(
+                    color: Color.fromARGB(255, 8, 46, 28),
+                    borderRadius: BorderRadius.all(
+                      Radius.circular(15),
+                    ),
+                  ),
+                  child: Center(
+                    child: Text(
+                      'salvar'.toUpperCase(),
+                      style: TextStyle(
+                          color: Colors.white, fontWeight: FontWeight.bold),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
