--- lib/views/inserirAplicacao.dart
@@ -1,394 +1,0 @@
-import 'package:flutter/material.dart';
-
-class inserirAplicacaoPage extends StatefulWidget {
-  const inserirAplicacaoPage({super.key});
-
-  @override
-  State<inserirAplicacaoPage> createState() => _inserirAplicacaoPageState();
-}
-
-@override
-class _inserirAplicacaoPageState extends State<inserirAplicacaoPage> {
-  String? _selectedValue_opcoesmedidas;
-  String? _selectedValue_medidatempo;
-
-  final List<String> _opcoes_medidas = [
-    '',
-    'L/ha',
-    'Kg/ha',
-    'mL/ha',
-    'g/ha',
-    'L/ac',
-    'Kg/ac',
-    'mL/L',
-    'g/L',
-    'UI/ha',
-    'gotas/m2',
-  ];
-
-  final List<String> _medidas_tempo = [
-    '',
-    'Horas',
-    'Dias',
-    'Semanas',
-    'Meses',
-    'Bimestres',
-    'Trimestres',
-    'Semestres',
-    'Anos',
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Row(
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: [
-            Text(
-              '<Nome da Lavoura>',
-              style: TextStyle(fontSize: 22),
-            ),
-          ],
-        ),
-        automaticallyImplyLeading: true,
-        flexibleSpace: Center(
-          child: Padding(
-            padding: const EdgeInsets.only(top: 30.0),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                Image.asset(
-                  'images/Logo_SafraApp3.png', // Substitua pelo caminho da sua imagem
-                  height: 40, // Ajuste a altura conforme necessário
-                ),
-              ],
-            ),
-          ),
-        ),
-        //automaticallyImplyLeading: false,
-        elevation: 0,
-        backgroundColor: Color.fromARGB(255, 2, 89, 47),
-        centerTitle: true,
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: [
-            Padding(
-              padding: const EdgeInsets.only(top: 28.0, left: 12),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Text(
-                    'Inserir Aplicação de Insumo',
-                    style: TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
-                  ),
-                ],
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 18.0),
-              child: Container(
-                width: MediaQuery.of(context).size.width / 1.1,
-                height: 50,
-                padding: EdgeInsets.all(8),
-                decoration: BoxDecoration(
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(50),
-                    ),
-                    color: Colors.white,
-                    border: Border.all(
-                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                child: TextField(
-                  //keyboardType: TextInputType.number,
-                  decoration: InputDecoration(
-                      border: InputBorder.none,
-                      hintText: 'Insumo',
-                      hintStyle:
-                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                ),
-              ),
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Padding(
-                  padding: const EdgeInsets.only(top: 18.0),
-                  child: Container(
-                    width: MediaQuery.of(context).size.width / 2.95,
-                    height: 50,
-                    padding: EdgeInsets.all(8),
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.all(
-                          Radius.circular(50),
-                        ),
-                        color: Colors.white,
-                        border: Border.all(
-                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                    child: TextField(
-                      keyboardType: TextInputType.datetime,
-                      decoration: InputDecoration(
-                          border: InputBorder.none,
-                          hintText: 'Data da Aplicação',
-                          hintStyle:
-                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                    ),
-                  ),
-                ),
-                SizedBox(width: 5.5),
-                Padding(
-                  padding: const EdgeInsets.only(top: 18.0),
-                  child: Container(
-                    width: MediaQuery.of(context).size.width / 3.25,
-                    height: 50,
-                    padding: EdgeInsets.all(8),
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.only(
-                          topLeft: Radius.circular(50),
-                          bottomLeft: Radius.circular(50),
-                        ),
-                        color: Colors.white,
-                        border: Border.all(
-                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                    child: TextField(
-                      keyboardType: TextInputType.number,
-                      decoration: InputDecoration(
-                          border: InputBorder.none,
-                          hintText: 'Dose Aplicada',
-                          hintStyle:
-                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.only(top: 18.0),
-                  child: Container(
-                    width: MediaQuery.of(context).size.width / 3.95,
-                    height: 50,
-                    padding: EdgeInsets.all(8),
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.only(
-                          topRight: Radius.circular(50),
-                          bottomRight: Radius.circular(50),
-                        ),
-                        color: Colors.white,
-                        border: Border.all(
-                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                    child: DropdownButton(
-                      value: _selectedValue_opcoesmedidas,
-                      isExpanded: true,
-                      hint: const Text(
-                        'Medida',
-                        style: TextStyle(
-                          color: Color.fromARGB(255, 8, 46, 28),
-                        ),
-                      ),
-                      items: _opcoes_medidas.map<DropdownMenuItem<String>>(
-                        (String value) {
-                          return DropdownMenuItem<String>(
-                            value: value,
-                            child: Text(value),
-                          );
-                        },
-                      ).toList(),
-                      onChanged: (String? newValue) {
-                        setState(() {
-                          _selectedValue_opcoesmedidas = newValue;
-                        });
-                      },
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 18.0),
-              child: Container(
-                width: MediaQuery.of(context).size.width / 1.1,
-                height: 50,
-                padding: EdgeInsets.all(8),
-                decoration: BoxDecoration(
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(50),
-                    ),
-                    color: Colors.white,
-                    border: Border.all(
-                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                child: TextField(
-                  //keyboardType: TextInputType.number,
-                  decoration: InputDecoration(
-                      border: InputBorder.none,
-                      hintText: 'Método de Aplicação',
-                      hintStyle:
-                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 18.0),
-              child: Container(
-                width: MediaQuery.of(context).size.width / 1.1,
-                height: 50,
-                padding: EdgeInsets.all(8),
-                decoration: BoxDecoration(
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(50),
-                    ),
-                    color: Colors.white,
-                    border: Border.all(
-                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                child: TextField(
-                  //keyboardType: TextInputType.number,
-                  decoration: InputDecoration(
-                      border: InputBorder.none,
-                      hintText: 'Objetivo da Aplicação',
-                      hintStyle:
-                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 18.0),
-              child: Container(
-                width: MediaQuery.of(context).size.width / 1.1,
-                height: 50,
-                padding: EdgeInsets.all(8),
-                decoration: BoxDecoration(
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(50),
-                    ),
-                    color: Colors.white,
-                    border: Border.all(
-                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                child: TextField(
-                  //keyboardType: TextInputType.number,
-                  decoration: InputDecoration(
-                      border: InputBorder.none,
-                      hintText: 'Responsável pela Aplicação',
-                      hintStyle:
-                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                ),
-              ),
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                SizedBox(width: 5.5),
-                Padding(
-                  padding: const EdgeInsets.only(top: 18.0),
-                  child: Container(
-                    width: MediaQuery.of(context).size.width / 1.52,
-                    height: 50,
-                    padding: EdgeInsets.all(8),
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.only(
-                          topLeft: Radius.circular(50),
-                          bottomLeft: Radius.circular(50),
-                        ),
-                        color: Colors.white,
-                        border: Border.all(
-                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                    child: TextField(
-                      keyboardType: TextInputType.number,
-                      decoration: InputDecoration(
-                          border: InputBorder.none,
-                          hintText: 'Quantidade do Período de Carência',
-                          hintStyle:
-                              TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: const EdgeInsets.only(top: 18.0),
-                  child: Container(
-                    width: MediaQuery.of(context).size.width / 3.95,
-                    height: 50,
-                    padding: EdgeInsets.all(8),
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.only(
-                          topRight: Radius.circular(50),
-                          bottomRight: Radius.circular(50),
-                        ),
-                        color: Colors.white,
-                        border: Border.all(
-                            color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                    child: DropdownButton(
-                      value: _selectedValue_medidatempo,
-                      isExpanded: true,
-                      hint: const Text(
-                        'Tempo',
-                        style: TextStyle(
-                          color: Color.fromARGB(255, 8, 46, 28),
-                        ),
-                      ),
-                      items: _medidas_tempo.map<DropdownMenuItem<String>>(
-                        (String value) {
-                          return DropdownMenuItem<String>(
-                            value: value,
-                            child: Text(value),
-                          );
-                        },
-                      ).toList(),
-                      onChanged: (String? newValue) {
-                        setState(() {
-                          _selectedValue_medidatempo = newValue;
-                        });
-                      },
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            Padding(
-              padding: const EdgeInsets.only(top: 18.0),
-              child: Container(
-                width: MediaQuery.of(context).size.width / 1.1,
-                height: 200,
-                padding: EdgeInsets.all(8),
-                decoration: BoxDecoration(
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(20),
-                    ),
-                    color: Colors.white,
-                    border: Border.all(
-                        color: Color.fromARGB(255, 8, 46, 28), width: 2.0)),
-                child: TextField(
-                  //keyboardType: TextInputType.number,
-                  decoration: InputDecoration(
-                      border: InputBorder.none,
-                      hintText:
-                          'Deixe aqui uma observação sobre esta aplicação',
-                      hintStyle:
-                          TextStyle(color: Color.fromARGB(255, 8, 46, 28))),
-                ),
-              ),
-            ),
-            Container(
-              padding: EdgeInsets.only(top: 50, bottom: 10),
-              child: InkWell(
-                onTap: () {},
-                child: Container(
-                  height: 60,
-                  width: MediaQuery.of(context).size.width / 1.1,
-                  decoration: BoxDecoration(
-                    color: Color.fromARGB(255, 8, 46, 28),
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(15),
-                    ),
-                  ),
-                  child: Center(
-                    child: Text(
-                      'salvar'.toUpperCase(),
-                      style: TextStyle(
-                          color: Colors.white, fontWeight: FontWeight.bold),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
